fix(tests): CryptoModulesTest 缺 import + Byte 字面量类型错误
- org.junit.Assert.assertFalse 未 import -> 新增 import
- 0xAA / 0xBB / 0xCC 在 Kotlin/JVM 中字面量默认 Int，赋给 Byte
  需要显式 .toByte()（之前只给 0xBB/0xCC 加了，漏掉 0xAA）

+++ app/src/test/java/com/mepass/app/CryptoModulesTest.kt
@@ -5,6 +5,7 @@ import com.mepass.app.crypto.Argon2Manager
 import com.mepass.app.crypto.ShamirSecretSharing
 import org.junit.Assert.assertArrayEquals
 import org.junit.Assert.assertEquals
+import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import kotlin.random.Random
@@ -70,7 +71,7 @@ class CryptoModulesTest {
 
     @Test
     fun `shamir k=1 n=5 all shares equal secret`() {
-        val secret = byteArrayOf(0xAA, 0xBB.toByte(), 0xCC.toByte())
+        val secret = byteArrayOf(0xAA.toByte(), 0xBB.toByte(), 0xCC.toByte())
         val shares = ShamirSecretSharing.split(secret, n = 5, k = 1)
         for (share in shares) {
             val recovered = ShamirSecretSharing.combine(listOf(share), 1)
